aider: Fix import statement in openai_assistant_manager.py to import the correct class from the openai package.
# Aider chat conversation:

USER: File "/Users/socrates/Development/chat-application/main.py", line 14, in <module>
    from api import router as api_router
  File "/Users/socrates/Development/chat-application/api.py", line 4, in <module>
    from openai_assistant_manager import OpenAIAssistantManager
  File "/Users/socrates/Development/chat-application/openai_assistant_manager.py", line 3, in <module>
    from openai import OpenAI_Client
ImportError: cannot import name 'OpenAI_Client' from 'openai' (/Users/socrates/Library/Caches/pypoetry/virtualenvs/chat-application-RBuUGsNS-py3.11/lib/python3.11/site-packages/openai/__init__.py)

ASSISTANT: The error message indicates that there is an attempt to import `OpenAI_Client` from the `openai` module, but such a class or function does not exist in the `openai` package. The correct class to import for interacting with the OpenAI API is `openai.OpenAI` or `openai.Client` depending on the version of the `openai` package you are using.

To fix this issue, we need to modify the `openai_assistant_manager.py` file to import the correct class from the `openai` package. Since I don't have the exact version of the `openai` package you are using, I will assume that the correct class to use is `openai.OpenAI` or `openai.Client`. If this is not the case, you may need to adjust the import statement according to the version of the `openai` package installed in your environment.

Here is the *SEARCH/REPLACE* block for the change:

openai_assistant_manager.py
<source>python
<<<<<<< SEARCH
from openai import OpenAI_Client
=======
from openai import OpenAI as OpenAI_Client
>>>>>>> REPLACE
</source>

Please replace `OpenAI as OpenAI_Client` with the correct class name if it's different in your installed version of the `openai` package. If you are unsure about the correct class to use, you can check the documentation of the `openai` package or the `__init__.py` file in the installed `openai` package directory to find the available classes.