aider: Update openai package to openai = "^1.2.3".
# Aider chat conversation:

USER: update openai package to openai = "^1.2.3"

ASSISTANT: To update the `openai` package version in the `pyproject.toml` file, you will need to modify the version constraint for the `openai` dependency. Here is the *SEARCH/REPLACE* block for the change:

pyproject.toml
<source>toml
<<<<<<< SEARCH
openai = "^0.10.2"
=======
openai = "^1.2.3"
>>>>>>> REPLACE
</source>

This change will update the version constraint for the `openai` package to `^1.2.3`, which means it will allow the installation of any version compatible with `1.2.3` but less than `2.0.0`. After making this change, you should run `poetry update` to update the package and lock file accordingly.